Allows `getDefaultAddress()` to return undefined.
Sometimes, you may be connected to an Ethereum node that doesn't have any accounts attached to it or exposed.  This should not prevent you from interfacing with the node in ways that don't require accounts such as doing local calls of methods that work with any account.

Also fixed a few styling issues in the output.