Adds support for overloads and duplicate contracts.
The handling of duplicate contracts and overloads is naive, the system just adds a number on the end of the contract/method name.  However, this solution is significantly better than not working.

Also fixes a bug with generated events which would result in non-compilable code if there were unexpected properties in the ABI.

Marked the parameters as readonly so it is more permissive in what it accepts.