Fixes a bug where tuple parameters couldn't be serialized.
Prior to this fix, the canonical signature was being passed to the call/send provider which meant that there were no parameter names present.  Since tuples are mapped to objects, this meant that the tuples could not be serialized since the generated code wanted an object for a tuple parameter rather than a list.  This is a major version bump because Dependency implementations potentially need to get the selector for a non-canonical function signature.

If you are using `@zoltu/ethereum-abi-encoder` this major version bump should be trivial as it already understands how to extract the selector from non-canonical signatures.