Updates to be compatible with NodeJS 14.
`output/index.js` is not a valid exports path, it has to be `./output/index.js`.

While I was in here I added support for targeting ES Modules for the library, since NodeJS 14 now supports that.  Also added a missing readme to `browser-dependencies` and cleaned up some issues with the tsconfig files.